feat: add command line interface to manage setup process
This commit introduces a command line interface (CLI) which adds more control to the setup process. It uses the 'clap' crate, a popular Rust package for creating CLIs. With this a user can now force the setup as needed by passing the 'force' flag during command. In this way, the setup process is not run every time the main function is run, saving unnecessary executions and system resources.