docs: add issue templates for bug report and feature request
This commit adds two new issue templates for bug reports and feature requests. This is to help users and contributors creating new issues by providing them with a proper format and ensuring all the necessary information is included in the issue. It would also help in automatically adding labels and assigning the issues to the right person. Furthermore, it disables the creation of blank issues.