refactor: update Prop Lot Query struct and GraphQL query for simplicity
Changes made include renaming the Query struct in fetcher.rs to IdeaQuery for better clarity. The structure of the GraphQL query in prop_lot_query.graphql was simplified by creating an IdeaFragment which encapsulates the main data we fetch when querying for ideas. This was done to improve the code readability and maintainability.