refactor: add comments fetching in Prop Lot module
This commit adds the ability to fetch comments using GraphQL in the prop_lot module. A new GraphQL query for fetching comments has been added in `prop_lot_query.graphql`. Also, the `fetch_comments` function, which is responsible for fetching these comments, has been added to the `fetcher.rs` file. This is to allow for comments data retrieval, further enhancing the functionality of the prop_lot module.