refactor: remove unnecessary clone method and handle string conversion
Changes are made to fetcher.rs, cacher.rs and mod.rs files in both prop_house and prop_lot directories, as well as handler.rs in prop_house directory. Unnecessary cloning has been removed to clean up the code and improve efficiency. In the fetcher.rs, the clone() method in the map function was replaced with direct value usage, reducing redundancy and complexity. Similarly, the cacher.rs files no longer use unnecessary clone where the value is already a reference. In the mod.rs files of both prop_house and prop_lot, the dereferencing was removed as Rust automatically dereferences arguments to methods and functions. Also, in handler.rs, we swapped a string for a character when replacing spaces in a title to lowercase as this is more efficient in Rust.