refactor: update code to handle new proposals and votes
Updated the structs Proposal and Vote to derive from the Debug trait to better track and debug data. Introduced handler functions for new instances of proposals and votes in both Mod.rs and Handler.rs. This update in handling responses will enable the application to respond more appropriately to the fetching of new proposals and votes.