refactor: add proposal title and link votes to proposals for Prop House
In this commit, additional fields were added to the `Proposal` and `Vote` structs. For better information traceability, the `Proposal` struct now includes the 'title' of the proposal. To link 'Vote' to a specific 'Proposal', 'proposal_id' was added to the `Vote` struct. With these changes, it is easier to track votes for a specific proposal and debug issues related to proposals and votes.