refactor: update caching process and improve logging
The caching mechanism was updated to allow batch setting of cache, which increases the efficiency by reducing the number of calls. This is especially beneficial when setting up the initial cache for ideas, votes, and comments. Instead of one by one, we now cache all the entities in one go.

Moreover, redundant comments in the `cacher.rs` file were removed, and error checking was added when handling new votes and comments in the `handler.rs` file.

Lastly, the `idea_id` field was added in the `Comment` struct to link comments to their respective ideas.

These changes will enhance the readability and performance of the code.