refactor: update vote and comment Prop Lot handler methods
Removed debug and error logs in the vote and comment handling methods and replaced them with the generation and execution of webhooks. This was done to provide a more complete notification mechanism. The idea is fetched from the cache and an error is raised if it doesn't exist. The vote/comment information is then used to construct an "Embed" object which is send via a webhook, giving a direct, link-filled notification each time a vote/comments comes in.