chore: add `package.json` for LilNouns bots project
This commit introduces a package.json file for the lilnouns-bots project. This file is essential as it holds various metadata relevant to the project such as its dependencies and scripts. It is essential for Node.js projects and it's being introduced to support the development and deployment phases. The file is marked as private to prevent accidental publication on npm. The wrangler module version 3.8.0 is added under devDependencies to help managing and publishing cloudflare workers.