ci: add PNPM dependencies maintenance to dependabot
This commit includes an additional package ecosystem, namely PNPM, to the dependabot configuration. The new rules are set to check for updates daily, automatically assign 'nekofar' and include the scope of the updates in the commit message. Pull requests are to be labelled as 'dependencies', allowing up to 20 open PRs at a time, and are raised against the 'develop' branch. This modification improves the maintenance of PNPM dependencies in our project.