refactor: update codebase to use PropHouse and PropLot structs
The codebase has been refactored to centralize data and logic related to PropHouse and PropLot into respective structs. Created a simple API for managing data related to PropHouse and PropLot, including methods to perform initial setup and subsequent updates. The change simplifies the process of bringing up and maintaining PropHouse and PropLot as now we just need to create instances of these structs and call their methods. This will simplify further development and maintenance.