refactor: revise graphql queries and fetcher for prop house
The commit is a revision to 'graphql/queries/prop_house_query.graphql' and 'src/prop_house/fetcher.rs'. In the graphql file, it modifies the 'AuctionFragment' from under 'createdDate' to under 'auctions'. For 'fetcher.rs', it retrieves and maps the 'auction_id' and 'proposal' from 'auctions' before populating the 'Proposal' struct.