refactor: update handlers to improve user-friendly naming
Replaced the get_ens method with get_domain_name, get_explorer_address, and get_short_address in both prop_lot and prop_house handler files. This change was done to improve the user-friendly display of addresses in the Discord embeds. Instead of displaying raw addresses or ENS hashes, it now shows a short version of the address or interpretable domain name from ENS, along with a direct link to the user's wallet in etherscan.