refactor: update visibility of Idea, Vote, Comment struct and their methods
The struct 'Idea', 'Vote', and 'Comment' in 'fetcher.rs' along with their respective fetch methods have been updated from 'pub(crate)' visibility to 'pub'. This change was made to allow these structs and methods to be accessible outside of the crate. They were previously confined to within the crate, but now they are exposed publicly and can be used in other parts of the project where needed.