refactor: implement Handler trait for DiscordHandler
The changes introduced include implementation of a new Handler trait for DiscordHandler. This was done to abstract the methods for handling new auctions, proposals and votes, which can now be thought of as implementations of a more abstract concept rather than functions tied to the DiscordHandler struct. This change should increase the code's maintainability and allow for easier future extensions. A marker trait has also been imported which provided functionality for boxing asynchronous operations.