refactor: update handlers to support multiple types
The code has been refactored to support multiple types of handlers. Previously, it was restricted to only a DiscordHandler. However, a need arose to add another handler and hence the code was adjusted to allow any number of handlers by storing them in a vector. This change allows for flexibility in adding or removing handlers without requiring code changes. Now, the handling of new auctions, votes, and proposals is done by all handlers in the vector. The commit also includes the addition of FarcasterHandler to the handlers vector.