style: refactor code to improve readability
This commit refactored several source code files to significantly improve the readability of the codes. In each file, there was a respace of imports from using multiple lines to a single line with curly brackets. Also, there was a movement from using minus signs to plus signs for indentation purpose. This gives a clearer understanding and a better view of the code structure.