refactor: enable channel key for Farcaster handlers
The "channelKey" has been uncommented in the JSON payloads of HTTP requests found in the farcaster handlers for both 'prop_house' and 'prop_lot'. Enabling "channelKey" ensures that these requests are directed to the desired channel, specifically "lil-nouns". Previously, the lack of channel direction could have caused requests to be misrouted.