refactor: replace wallet address fetching method
Replaced the method for fetching wallet addresses from get_domain_name to get_wallet_holder in the FarcasterHandler of prop_house, meta_gov, and prop_lot modules.

The change was made to fetch the ENS associated handle, if exists, instead of the domain name for the user's wallet. This makes the user's representation in the application more meaningful and personalized. In case there's no handle associated, it falls back to the previous implementation of showing a short version of the address.